POC/Clustering/requirements.txt
scikit-learn
flask-cors
pandas
matplotlib
pyjwt